requirements.txt
Flask==3.0.0
redis==5.0.1
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Werkzeug==3.0.1
itsdangerous==2.1.2
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.23